requirements/base.txt
-i https://pypi.org/simple

amqp==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
cx-oracle==7.3.0
django-cas-ng==4.1.1
django-celery-beat==2.0.0
django-environ==0.4.5
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-redis==4.12.1
django-timezone-field==4.0; python_version >= '3.5'
django==2.2.14
djangorestframework==3.11.0
drf-nested-routers==0.91
factory-boy==2.12.0
faker==4.1.1; python_version >= '3.4'
flower==0.9.5
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
humanize==2.5.0; python_version >= '3.5'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.7.0; python_version < '3.8'
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
lxml==4.5.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
prometheus-client==0.8.0
psycopg2-binary==2.8.5
python-cas==1.5.0
python-crontab==2.5.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests==2.24.0
sentry-sdk==0.16.2
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
text-unidecode==1.3
tornado==6.0.4; python_full_version >= '3.5.2'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
zipp==3.1.0; python_version >= '3.6'
